General(:zap:): Abstracted the menu.
