+++ src/components/Menu/index.ts
@@ -1,0 +1,3 @@
+import Menu from './Menu';
+
+export default Menu;
