+++ src/pages/Home/Home.tsx
@@ -5,18 +5,8 @@ import Pages from '../../enum/Pages';
 import {ProjectProps, projects} from '../../utils/DummyData';
 import Button from '../../components/Button';
 import {bgcolor, color} from '../../utils/DarkmodeUtils';
-import Animated, {
-  useAnimatedScrollHandler,
-  useAnimatedStyle,
-  useSharedValue,
-  withRepeat,
-  withTiming,
-} from 'react-native-reanimated';
-import CustomIcon from '../../components/Icon';
-import Icons from '../../enum/Icons';
-import {colors, metrics} from '../../themes';
-import useMenu from './animations/useMenu';
-import {isAndroid} from '../../utils/PlatformUtils';
+import Animated from 'react-native-reanimated';
+import Menu from '../../components/Menu';
 
 interface Props {
   darkmode: boolean;
@@ -26,61 +16,9 @@ interface Props {
 const AnimatedScrollview = Animated.createAnimatedComponent(ScrollView);
 
 const Home = ({navigation, darkmode}: Props) => {
-  const [menuActive, setMenuActive] = useState(false);
-
-  const toggleMenu = useCallback(
-    (isActive: boolean) => {
-      setMenuActive(isActive);
-    },
-    [setMenuActive],
-  );
-
-  const {
-    methods: {startMenuAnimation, closeMenu},
-    styles: {scaleStyle, opacityStyle},
-  } = useMenu(toggleMenu);
-
   return (
     <View style={styles.container}>
-      <Animated.View
-        style={[
-          {
-            ...StyleSheet.absoluteFillObject,
-            backgroundColor: colors.black,
-          },
-          menuActive && {zIndex: 1},
-          opacityStyle,
-        ]}>
-        <Button
-          onPress={() => navigation.navigate(Pages.SETTINGS)}
-          style={styles.menuItem}>
-          <Text style={[color(darkmode), styles.menuItemText]}>Settings</Text>
-        </Button>
-        <Button
-          onPress={() => navigation.navigate(Pages.CPHILIPSE)}
-          style={styles.menuItem2}>
-          <Text style={[color(darkmode), styles.menuItemText]}>
-            The creator
-          </Text>
-        </Button>
-      </Animated.View>
-      <View style={styles.menuContainer}>
-        <Button
-          onPress={menuActive ? closeMenu : startMenuAnimation}
-          style={styles.menu}>
-          {menuActive ? (
-            <View style={styles.closeIcon}>
-              <CustomIcon name={Icons.CLOSE} color={colors.white} />
-            </View>
-          ) : (
-            <>
-              <View style={styles.lineOne} />
-              <View style={styles.lineTwo} />
-            </>
-          )}
-        </Button>
-      </View>
-      <Animated.View style={[styles.menuCircle, scaleStyle]} />
+      <Menu darkmode={darkmode} goToPage={navigation.navigate} />
 
       <Button
         onPress={() => navigation.navigate(Pages.CPHILIPSE)}
